split flex class to a separate file
